Add a `make test` target
Previous coredns versions had a `make test` target and our ci-operator
currently uses this target to kick off the tests. So, at least to get
tests running against this PR without requiring changes to the
openshift/release ci-operator setup, I'm adding that target back.

This just runs the full litany of tests that Travis would run except
for the coverage targets and the tests that require a running etcd
server.